自定义协议方式完成

+++ service-member-api/src/main/java/com.crown.dmember/MemberService.java
@@ -3,5 +3,5 @@ package com.crown.dmember;
 import com.crown.servicecommon.domain.UserInfo;
 
 public interface MemberService {
-    UserInfo getUser(Long id);
+    UserInfo getUser(int id);
 }
